Version 5.13.11.7 CODE QUALITY - COMPLETE WARNING ELIMINATION: Fixed ALL remaining compiler warnings for perfectly clean build! *					C++ WARNINGS FIXED: C4018 in RestoreEngine_Advanced.cpp line 740 - signed/unsigned mismatch in loop comparison *					(int i < DWORD wimInfo.ImageCount). Changed loop variable from 'int i' to 'DWORD i' and added explicit cast *					'imageIndex + static_cast<int>(i)' for safe signed/unsigned arithmetic. WIM API uses DWORD for image counts *					(unsigned 32-bit) requiring careful type management. C# WARNINGS FIXED: 1) CS4014 (unawaited async calls) - 4 instances *					fixed in ServiceManagementWindow.xaml.cs and MainWindow.xaml.cs by adding discard pattern '_ =' to fire-and-forget *					async operations (version checking, error dialogs). Indicates intentional non-blocking behavior. 2) CS8600 (null *					literal conversion) - fixed in TempPathSelectionDialog.xaml.cs by changing 'string root' to 'string? root' to *					properly handle nullable return from Path.GetPathRoot(). 3) CS8602 (null reference dereference) - 2 instances fixed *					in JobManager.cs lines 152 & 159 by adding null-forgiving operator '!' to 'job.Schedule!.NextRunTime' since Schedule *					is guaranteed non-null after CalculateNextRunTime() call. RESULTS: Build now completes with 0 errors, 0 warnings across *					all 3 projects (BackupEngine C++, BackupService C#, BackupUI C#)! Clean build ensures: No hidden bugs from type mismatches, *					Proper null safety throughout, Clear code intent with explicit patterns, Compiler optimizations work correctly, Future *					warnings immediately visible. Production-ready warning-free codebase following modern C++ and C# best practices! *					Enterprise-grade code quality with zero technical debt from compiler warnings! Perfect foundation for continued *					development without warning noise masking real issues! mdail 3/12/2026

+++ BackupEngine/RestoreEngine_Advanced.cpp
@@ -737,8 +737,8 @@ extern "C" {
             // For now, restore starting from the specified image index
             // (UI should select the first image of the backup set)
 
-            for (int i = 0; i < wimInfo.ImageCount; i++) {
-                int currentIndex = imageIndex + i;
+            for (DWORD i = 0; i < wimInfo.ImageCount; i++) {
+                int currentIndex = imageIndex + static_cast<int>(i);
                 if (currentIndex > static_cast<int>(wimInfo.ImageCount)) break;
 
                 HANDLE hImage = WIMLoadImage(hWim, currentIndex);
